Fallback to merging accounts in reverse

+++ src/auth.rs
@@ -559,31 +559,36 @@ pub(crate) enum MergeAccountsError {
 #[rocket::get("/merge-accounts")]
 pub(crate) async fn merge_accounts(pool: &State<PgPool>, me: User, racetime_user: Option<RaceTimeUser>, discord_user: Option<DiscordUser>) -> Result<Redirect, MergeAccountsError> {
     let mut transaction = pool.begin().await?;
-    match (me.racetime, me.discord) {
-        (Some(_), Some(_)) => return Err(MergeAccountsError::AlreadyMerged),
-        (Some(_), None) => if let Some(discord_user) = discord_user {
-            if let Ok(Some(discord_user)) = User::from_discord(&mut *transaction, discord_user.id).await {
-                if discord_user.racetime.is_none() {
-                    let discord = discord_user.discord.expect("Discord user without Discord ID");
-                    sqlx::query!("DELETE FROM users WHERE id = $1", discord_user.id as _).execute(&mut *transaction).await?;
-                    sqlx::query!("UPDATE users SET discord_id = $1, discord_display_name = $2, discord_discriminator = $3, discord_username = $4 WHERE id = $5", PgSnowflake(discord.id) as _, discord.display_name, discord.username_or_discriminator.as_ref().right() as _, discord.username_or_discriminator.as_ref().left(), me.id as _).execute(&mut *transaction).await?;
-                    transaction.commit().await?;
-                    return Ok(Redirect::to(uri!(crate::user::profile(me.id))))
+    match (me.racetime.is_some(), me.discord.is_some()) {
+        (false, false) => unreachable!("signed in but neither account connected"),
+        (true, true) => return Err(MergeAccountsError::AlreadyMerged),
+        (is_racetime, _) => if_chain! {
+            if is_racetime;
+            if let Some(discord_user) = discord_user;
+            if let Ok(Some(discord_user)) = User::from_discord(&mut *transaction, discord_user.id).await;
+            if discord_user.racetime.is_none();
+            let discord = discord_user.discord.expect("Discord user without Discord ID");
+            if sqlx::query!("DELETE FROM users WHERE id = $1", discord_user.id as _).execute(&mut *transaction).await.is_ok();
+            if sqlx::query!("UPDATE users SET discord_id = $1, discord_display_name = $2, discord_discriminator = $3, discord_username = $4 WHERE id = $5", PgSnowflake(discord.id) as _, discord.display_name, discord.username_or_discriminator.as_ref().right() as _, discord.username_or_discriminator.as_ref().left(), me.id as _).execute(&mut *transaction).await.is_ok();
+            then {
+                transaction.commit().await?;
+                return Ok(Redirect::to(uri!(crate::user::profile(me.id))))
+            } else {
+                transaction.rollback().await?;
+                transaction = pool.begin().await?;
+                if let Some(racetime_user) = racetime_user {
+                    if let Ok(Some(racetime_user)) = User::from_racetime(&mut *transaction, &racetime_user.id).await {
+                        if racetime_user.discord.is_none() {
+                            let racetime = racetime_user.racetime.expect("racetime.gg user without racetime.gg ID");
+                            sqlx::query!("DELETE FROM users WHERE id = $1", racetime_user.id as _).execute(&mut *transaction).await?;
+                            sqlx::query!("UPDATE users SET racetime_id = $1, racetime_display_name = $2, racetime_discriminator = $3, racetime_pronouns = $4 WHERE id = $5", racetime.id, racetime.display_name, racetime.discriminator as _, racetime.pronouns as _, me.id as _).execute(&mut *transaction).await?;
+                            transaction.commit().await?;
+                            return Ok(Redirect::to(uri!(crate::user::profile(me.id))))
+                        }
+                    }
                 }
             }
         },
-        (None, Some(_)) => if let Some(racetime_user) = racetime_user {
-            if let Ok(Some(racetime_user)) = User::from_racetime(&mut *transaction, &racetime_user.id).await {
-                if racetime_user.discord.is_none() {
-                    let racetime = racetime_user.racetime.expect("racetime.gg user without racetime.gg ID");
-                    sqlx::query!("DELETE FROM users WHERE id = $1", racetime_user.id as _).execute(&mut *transaction).await?;
-                    sqlx::query!("UPDATE users SET racetime_id = $1, racetime_display_name = $2, racetime_discriminator = $3, racetime_pronouns = $4 WHERE id = $5", racetime.id, racetime.display_name, racetime.discriminator as _, racetime.pronouns as _, me.id as _).execute(&mut *transaction).await?;
-                    transaction.commit().await?;
-                    return Ok(Redirect::to(uri!(crate::user::profile(me.id))))
-                }
-            }
-        },
-        (None, None) => unreachable!("signed in but neither account connected"),
     }
     transaction.rollback().await?;
     Err(MergeAccountsError::Other)
